flashinspector-ai/requirements.txt
ultralytics>=8.0.0
roboflow>=1.1.0
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0
python-dotenv>=1.0.0
pyyaml>=6.0
numpy>=1.24.0
pillow>=10.0.0
tqdm>=4.65.0
